refactor(backup): split BackupRestore into focused modules
Extract backup table, dialogs, types, constants, and formatting helpers into dedicated files to reduce component complexity and keep the main BackupRestore component as a clear orchestration layer.
